feat: api-handlers-decomposition-recovery-002 - Extract work route handlers
