+++ pkg/api/handlers.go
@@ -5,10 +5,8 @@ package api
 import (
 	"bytes"
 	"context"
-	"encoding/base64"
 	"encoding/json"
 	"errors"
-	"fmt"
 	"io"
 	"net/http"
 	"sort"
@@ -18,18 +16,14 @@ import (
 	"github.com/portpowered/infinite-you/pkg/apisurface"
 	"github.com/portpowered/infinite-you/pkg/apisurface/optional"
 	factoryconfig "github.com/portpowered/infinite-you/pkg/config"
-	factoryrequests "github.com/portpowered/infinite-you/pkg/factory/requests"
-	factoryvalidation "github.com/portpowered/infinite-you/pkg/factory/validation"
 	"github.com/portpowered/infinite-you/pkg/factory/state"
-	"github.com/portpowered/infinite-you/pkg/factorysessions"
+	factoryvalidation "github.com/portpowered/infinite-you/pkg/factory/validation"
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 	"github.com/portpowered/infinite-you/pkg/petri"
 	"github.com/portpowered/infinite-you/pkg/workcontent"
 	workerprompting "github.com/portpowered/infinite-you/pkg/workers/prompting"
 	"go.uber.org/zap"
 )
-
-const defaultMaxResults = 50
 
 var _ factoryapi.ServerInterface = (*Server)(nil)
 
@@ -41,250 +35,6 @@ func (s *Server) requireSessionRuntime(w http.ResponseWriter) (apisurface.Sessio
 		return nil, false
 	}
 	return s.sessionRuntime, true
-}
-
-func (s *Server) SubmitWork(w http.ResponseWriter, r *http.Request) {
-	req, err := decodeSubmitWorkRequestBody(r.Body)
-	if err != nil {
-		if message, ok := requestFieldValidationMessage(err); ok {
-			s.writeError(w, http.StatusBadRequest, message, "BAD_REQUEST")
-			return
-		}
-		s.writeError(w, http.StatusBadRequest, "invalid request payload", "BAD_REQUEST")
-		return
-	}
-
-	if req.WorkTypeName == "" {
-		s.writeError(w, http.StatusBadRequest, "workTypeName is required", "BAD_REQUEST")
-		return
-	}
-
-	payload, err := generatedPayloadToRawMessage(req.Payload)
-	if err != nil {
-		s.writeError(w, http.StatusBadRequest, "invalid request payload", "BAD_REQUEST")
-		return
-	}
-	content, err := submitWorkContent(req)
-	if err != nil {
-		if message, ok := requestFieldValidationMessage(err); ok {
-			s.writeError(w, http.StatusBadRequest, message, "BAD_REQUEST")
-			return
-		}
-		s.writeError(w, http.StatusBadRequest, "invalid request payload", "BAD_REQUEST")
-		return
-	}
-
-	submitReq := interfaces.SubmitRequest{
-		Name:                   strings.TrimSpace(req.Name),
-		WorkTypeID:             req.WorkTypeName,
-		CurrentChainingTraceID: stringValue(req.CurrentChainingTraceId),
-		TraceID:                factoryrequests.ResolveWorkRequestCurrentChainingTraceID(stringValue(req.CurrentChainingTraceId), stringValue(req.TraceId)),
-		Content:                content,
-		Payload:                payload,
-		Tags:                   generatedStringMap(req.Tags),
-		Relations:              generatedSubmitRelations(req.Relations),
-	}
-	workRequest := factoryrequests.WorkRequestFromSubmitRequests([]interfaces.SubmitRequest{submitReq})
-
-	result, err := s.runtime.SubmitWorkRequest(r.Context(), workRequest)
-	if err != nil {
-		if message, ok := submitWorkBadRequestMessage(err); ok {
-			s.writeError(w, http.StatusBadRequest, message, "BAD_REQUEST")
-			return
-		}
-		s.logger.Error("submit work failed", zap.Error(err))
-		s.writeError(w, http.StatusInternalServerError, "failed to submit work", "INTERNAL_ERROR")
-		return
-	}
-
-	s.writeJSON(w, http.StatusCreated, submitWorkResponseFromResult(result, factorysessions.DefaultSessionID))
-}
-
-func (s *Server) SubmitWorkBySessionId(w http.ResponseWriter, r *http.Request, sessionID factoryapi.SessionID) {
-	sessionRuntime, ok := s.requireSessionRuntime(w)
-	if !ok {
-		return
-	}
-
-	req, err := decodeSubmitWorkRequestBody(r.Body)
-	if err != nil {
-		if message, ok := requestFieldValidationMessage(err); ok {
-			s.writeError(w, http.StatusBadRequest, message, "BAD_REQUEST")
-			return
-		}
-		s.writeError(w, http.StatusBadRequest, "invalid request payload", "BAD_REQUEST")
-		return
-	}
-
-	if req.WorkTypeName == "" {
-		s.writeError(w, http.StatusBadRequest, "workTypeName is required", "BAD_REQUEST")
-		return
-	}
-
-	payload, err := generatedPayloadToRawMessage(req.Payload)
-	if err != nil {
-		s.writeError(w, http.StatusBadRequest, "invalid request payload", "BAD_REQUEST")
-		return
-	}
-	content, err := submitWorkContent(req)
-	if err != nil {
-		if message, ok := requestFieldValidationMessage(err); ok {
-			s.writeError(w, http.StatusBadRequest, message, "BAD_REQUEST")
-			return
-		}
-		s.writeError(w, http.StatusBadRequest, "invalid request payload", "BAD_REQUEST")
-		return
-	}
-
-	submitReq := interfaces.SubmitRequest{
-		Name:                   strings.TrimSpace(req.Name),
-		WorkTypeID:             req.WorkTypeName,
-		CurrentChainingTraceID: stringValue(req.CurrentChainingTraceId),
-		TraceID:                factoryrequests.ResolveWorkRequestCurrentChainingTraceID(stringValue(req.CurrentChainingTraceId), stringValue(req.TraceId)),
-		Content:                content,
-		Payload:                payload,
-		Tags:                   generatedStringMap(req.Tags),
-		Relations:              generatedSubmitRelations(req.Relations),
-	}
-	workRequest := factoryrequests.WorkRequestFromSubmitRequests([]interfaces.SubmitRequest{submitReq})
-
-	result, err := sessionRuntime.SubmitWorkRequestForSession(r.Context(), string(sessionID), workRequest)
-	if err != nil {
-		if errors.Is(err, apisurface.ErrFactorySessionNotFound) {
-			s.writeError(w, http.StatusNotFound, "factory session not found", "NOT_FOUND")
-			return
-		}
-		if message, ok := submitWorkBadRequestMessage(err); ok {
-			s.writeError(w, http.StatusBadRequest, message, "BAD_REQUEST")
-			return
-		}
-		s.logger.Error("submit work failed", zap.Error(err), zap.String("session_id", string(sessionID)))
-		s.writeError(w, http.StatusInternalServerError, "failed to submit work", "INTERNAL_ERROR")
-		return
-	}
-
-	s.writeJSON(w, http.StatusCreated, submitWorkResponseFromResult(result, string(sessionID)))
-}
-
-func (s *Server) UpsertWorkRequest(w http.ResponseWriter, r *http.Request, requestID string) {
-	req, err := decodeWorkRequestBody(r.Body)
-	if err != nil {
-		if message, ok := requestFieldValidationMessage(err); ok {
-			s.writeError(w, http.StatusBadRequest, message, "BAD_REQUEST")
-			return
-		}
-		s.writeError(w, http.StatusBadRequest, "invalid request payload", "BAD_REQUEST")
-		return
-	}
-
-	if requestID == "" {
-		s.writeError(w, http.StatusBadRequest, "request_id is required", "BAD_REQUEST")
-		return
-	}
-	if req.RequestId == "" {
-		s.writeError(w, http.StatusBadRequest, "requestId is required", "BAD_REQUEST")
-		return
-	}
-	if req.RequestId != requestID {
-		s.writeError(w, http.StatusBadRequest, "request_id path and requestId body must match", "BAD_REQUEST")
-		return
-	}
-
-	workRequest, err := generatedWorkRequestToDomain(req)
-	if err != nil {
-		if message, ok := requestFieldValidationMessage(err); ok {
-			s.writeError(w, http.StatusBadRequest, message, "BAD_REQUEST")
-			return
-		}
-		s.writeError(w, http.StatusBadRequest, "invalid request payload", "BAD_REQUEST")
-		return
-	}
-	applyStableTraceToWorkRequest(&workRequest)
-	result, err := s.runtime.SubmitWorkRequest(r.Context(), workRequest)
-	if err != nil {
-		if strings.HasPrefix(err.Error(), "work_request:") {
-			s.writeError(w, http.StatusBadRequest, submitWorkTypeNameMessage(err.Error()), "BAD_REQUEST")
-			return
-		}
-		s.logger.Error("upsert work request failed", zap.Error(err))
-		s.writeError(w, http.StatusInternalServerError, "failed to submit work request", "INTERNAL_ERROR")
-		return
-	}
-
-	s.writeJSON(w, http.StatusCreated, upsertWorkRequestResponse(result))
-}
-
-func (s *Server) UpsertWorkRequestBySessionId(w http.ResponseWriter, r *http.Request, sessionID factoryapi.SessionID, requestID string) {
-	sessionRuntime, ok := s.requireSessionRuntime(w)
-	if !ok {
-		return
-	}
-
-	req, err := decodeWorkRequestBody(r.Body)
-	if err != nil {
-		if message, ok := requestFieldValidationMessage(err); ok {
-			s.writeError(w, http.StatusBadRequest, message, "BAD_REQUEST")
-			return
-		}
-		s.writeError(w, http.StatusBadRequest, "invalid request payload", "BAD_REQUEST")
-		return
-	}
-
-	if requestID == "" {
-		s.writeError(w, http.StatusBadRequest, "request_id is required", "BAD_REQUEST")
-		return
-	}
-	if req.RequestId == "" {
-		s.writeError(w, http.StatusBadRequest, "requestId is required", "BAD_REQUEST")
-		return
-	}
-	if req.RequestId != requestID {
-		s.writeError(w, http.StatusBadRequest, "request_id path and requestId body must match", "BAD_REQUEST")
-		return
-	}
-
-	workRequest, err := generatedWorkRequestToDomain(req)
-	if err != nil {
-		if message, ok := requestFieldValidationMessage(err); ok {
-			s.writeError(w, http.StatusBadRequest, message, "BAD_REQUEST")
-			return
-		}
-		s.writeError(w, http.StatusBadRequest, "invalid request payload", "BAD_REQUEST")
-		return
-	}
-	applyStableTraceToWorkRequest(&workRequest)
-	result, err := sessionRuntime.SubmitWorkRequestForSession(r.Context(), string(sessionID), workRequest)
-	if err != nil {
-		if errors.Is(err, apisurface.ErrFactorySessionNotFound) {
-			s.writeError(w, http.StatusNotFound, "factory session not found", "NOT_FOUND")
-			return
-		}
-		if strings.HasPrefix(err.Error(), "work_request:") {
-			s.writeError(w, http.StatusBadRequest, submitWorkTypeNameMessage(err.Error()), "BAD_REQUEST")
-			return
-		}
-		s.logger.Error("upsert work request failed", zap.Error(err), zap.String("session_id", string(sessionID)))
-		s.writeError(w, http.StatusInternalServerError, "failed to submit work request", "INTERNAL_ERROR")
-		return
-	}
-
-	s.writeJSON(w, http.StatusCreated, upsertWorkRequestResponse(result))
-}
-
-func upsertWorkRequestResponse(result interfaces.WorkRequestSubmitResult) factoryapi.UpsertWorkRequestResponse {
-	works := make([]factoryapi.UpsertWorkRequestSubmittedWork, 0, len(result.Works))
-	for _, work := range result.Works {
-		works = append(works, factoryapi.UpsertWorkRequestSubmittedWork{
-			Name:         work.Name,
-			WorkTypeName: work.WorkTypeName,
-			WorkId:       work.WorkID,
-		})
-	}
-	return factoryapi.UpsertWorkRequestResponse{
-		RequestId: result.RequestID,
-		TraceId:   result.TraceID,
-		Works:     works,
-	}
 }
 
 func (s *Server) ValidateFactory(w http.ResponseWriter, r *http.Request) {
@@ -682,319 +432,6 @@ func (s *Server) writeCurrentFactoryError(
 	}
 }
 
-func (s *Server) ListWork(w http.ResponseWriter, r *http.Request, params factoryapi.ListWorkParams) {
-	s.listWork(w, r, params, s.runtime.GetEngineStateSnapshot)
-}
-
-func (s *Server) ListWorkBySessionId(w http.ResponseWriter, r *http.Request, sessionID factoryapi.SessionID, params factoryapi.ListWorkBySessionIdParams) {
-	sessionRuntime, ok := s.requireSessionRuntime(w)
-	if !ok {
-		return
-	}
-	legacyParams := factoryapi.ListWorkParams{
-		MaxResults:   params.MaxResults,
-		NextToken:    params.NextToken,
-		StateName:    params.StateName,
-		Name:         params.Name,
-		WorkTypeName: params.WorkTypeName,
-		TraceId:      params.TraceId,
-	}
-	if params.StateType != nil {
-		stateType := factoryapi.WorkStateType(*params.StateType)
-		legacyParams.StateType = &stateType
-	}
-	if params.SortBy != nil {
-		sortBy := factoryapi.ListWorkParamsSortBy(*params.SortBy)
-		legacyParams.SortBy = &sortBy
-	}
-	s.listWork(w, r, legacyParams, func(ctx context.Context) (*interfaces.EngineStateSnapshot[petri.MarkingSnapshot, *state.Net], error) {
-		return sessionRuntime.GetEngineStateSnapshotForSession(ctx, string(sessionID))
-	})
-}
-
-func (s *Server) listWork(
-	w http.ResponseWriter,
-	r *http.Request,
-	params factoryapi.ListWorkParams,
-	loadSnapshot func(context.Context) (*interfaces.EngineStateSnapshot[petri.MarkingSnapshot, *state.Net], error),
-) {
-	if params.StateType != nil && !validWorkStateType(factoryapi.WorkStateType(*params.StateType)) {
-		s.writeError(w, http.StatusBadRequest, "state.type must be one of INITIAL, PROCESSING, TERMINAL, or FAILED", "BAD_REQUEST")
-		return
-	}
-	if params.SortBy != nil && *params.SortBy != factoryapi.ListWorkParamsSortByStateType {
-		s.writeError(w, http.StatusBadRequest, "sortBy must be state.type", "BAD_REQUEST")
-		return
-	}
-
-	snapshot, err := loadSnapshot(r.Context())
-	if err != nil {
-		if errors.Is(err, apisurface.ErrFactorySessionNotFound) {
-			s.writeError(w, http.StatusNotFound, "factory session not found", "NOT_FOUND")
-			return
-		}
-		s.logger.Error("get engine state snapshot failed", zap.Error(err))
-		s.writeError(w, http.StatusInternalServerError, "failed to get engine state snapshot", "INTERNAL_ERROR")
-		return
-	}
-
-	// Collect, filter, and sort public work for deterministic pagination.
-	workNamesByID := publicWorkNamesByID(snapshot.Marking.Tokens)
-	items := make([]listWorkItem, 0, len(snapshot.Marking.Tokens))
-	for _, t := range snapshot.Marking.Tokens {
-		if !publicWorkToken(t) {
-			continue
-		}
-		work := tokenToWork(t, snapshot.Topology)
-		work.Relations = generatedWorkRelations(t, work.Name, workNamesByID)
-		if !workMatchesListFilters(work, params) {
-			continue
-		}
-		items = append(items, listWorkItem{cursorID: t.ID, work: work})
-	}
-	sortListWorkItems(items, listWorkSortMode(params.SortBy))
-
-	// Consume the generated route params directly. Non-positive values still fall back
-	// to the default page size after successful integer binding.
-	maxResults := defaultMaxResults
-	if params.MaxResults != nil && *params.MaxResults > 0 {
-		maxResults = *params.MaxResults
-	}
-
-	startIdx := 0
-	if cursor := stringValue(params.NextToken); cursor != "" {
-		decoded, err := base64.StdEncoding.DecodeString(cursor)
-		if err == nil {
-			startIdx = nextListWorkIndex(items, string(decoded))
-		}
-	}
-
-	// Slice the results.
-	end := min(startIdx+maxResults, len(items))
-	page := items[startIdx:end]
-
-	resp := factoryapi.ListWorkResponse{
-		Results: listWorkResults(page),
-		PaginationContext: &factoryapi.PaginationContext{
-			MaxResults: maxResults,
-		},
-	}
-	if end < len(items) {
-		lastID := page[len(page)-1].cursorID
-		nextToken := base64.StdEncoding.EncodeToString([]byte(lastID))
-		resp.PaginationContext.NextToken = &nextToken
-	}
-
-	s.writeJSON(w, http.StatusOK, resp)
-}
-
-func validWorkStateType(stateType factoryapi.WorkStateType) bool {
-	switch stateType {
-	case factoryapi.WorkStateTypeINITIAL,
-		factoryapi.WorkStateTypePROCESSING,
-		factoryapi.WorkStateTypeTERMINAL,
-		factoryapi.WorkStateTypeFAILED:
-		return true
-	default:
-		return false
-	}
-}
-
-type listWorkItem struct {
-	cursorID string
-	work     factoryapi.Work
-}
-
-type listWorkSortModeValue int
-
-const (
-	listWorkSortDefault listWorkSortModeValue = iota
-	listWorkSortStateType
-)
-
-func listWorkSortMode(sortBy *factoryapi.ListWorkParamsSortBy) listWorkSortModeValue {
-	if sortBy != nil && *sortBy == factoryapi.ListWorkParamsSortByStateType {
-		return listWorkSortStateType
-	}
-	return listWorkSortDefault
-}
-
-func sortListWorkItems(items []listWorkItem, mode listWorkSortModeValue) {
-	sort.Slice(items, func(i, j int) bool {
-		left := items[i]
-		right := items[j]
-		if mode == listWorkSortStateType {
-			return lessListWorkByStateType(left, right)
-		}
-
-		leftOrder := listWorkStateOrder(left.work.State)
-		rightOrder := listWorkStateOrder(right.work.State)
-		if leftOrder != rightOrder {
-			return leftOrder < rightOrder
-		}
-
-		leftStateType := listWorkStateType(left.work.State)
-		rightStateType := listWorkStateType(right.work.State)
-		if leftStateType != rightStateType {
-			return leftStateType < rightStateType
-		}
-
-		return left.cursorID < right.cursorID
-	})
-}
-
-func lessListWorkByStateType(left, right listWorkItem) bool {
-	leftStateType := listWorkStateType(left.work.State)
-	rightStateType := listWorkStateType(right.work.State)
-	if leftStateType != rightStateType {
-		return leftStateType < rightStateType
-	}
-	return left.cursorID < right.cursorID
-}
-
-func listWorkStateOrder(workState *factoryapi.WorkState) int {
-	if workState == nil {
-		return 4
-	}
-	switch workState.Type {
-	case factoryapi.WorkStateTypeINITIAL:
-		return 0
-	case factoryapi.WorkStateTypePROCESSING:
-		return 1
-	case factoryapi.WorkStateTypeFAILED:
-		return 2
-	case factoryapi.WorkStateTypeTERMINAL:
-		return 3
-	default:
-		return 4
-	}
-}
-
-func listWorkStateType(workState *factoryapi.WorkState) string {
-	if workState == nil {
-		return ""
-	}
-	return string(workState.Type)
-}
-
-func nextListWorkIndex(items []listWorkItem, cursorID string) int {
-	for i, item := range items {
-		if item.cursorID == cursorID {
-			return i + 1
-		}
-	}
-	return len(items)
-}
-
-func listWorkResults(items []listWorkItem) []factoryapi.Work {
-	results := make([]factoryapi.Work, len(items))
-	for i, item := range items {
-		results[i] = item.work
-	}
-	return results
-}
-
-func workMatchesListFilters(work factoryapi.Work, params factoryapi.ListWorkParams) bool {
-	return workMatchesStateListFilters(work, params) &&
-		workMatchesNameListFilter(work, params) &&
-		workMatchesWorkTypeNameListFilter(work, params) &&
-		workMatchesTraceIDListFilter(work, params)
-}
-
-func workMatchesStateListFilters(work factoryapi.Work, params factoryapi.ListWorkParams) bool {
-	if params.StateName != nil {
-		if work.State == nil || work.State.Name != *params.StateName {
-			return false
-		}
-	}
-	if params.StateType != nil {
-		if work.State == nil || work.State.Type != *params.StateType {
-			return false
-		}
-	}
-	return true
-}
-
-func workMatchesNameListFilter(work factoryapi.Work, params factoryapi.ListWorkParams) bool {
-	if params.Name == nil || *params.Name == "" {
-		return true
-	}
-	return strings.Contains(strings.ToLower(work.Name), strings.ToLower(string(*params.Name)))
-}
-
-func workMatchesWorkTypeNameListFilter(work factoryapi.Work, params factoryapi.ListWorkParams) bool {
-	if params.WorkTypeName == nil || *params.WorkTypeName == "" {
-		return true
-	}
-	return stringValue(work.WorkTypeName) == string(*params.WorkTypeName)
-}
-
-func workMatchesTraceIDListFilter(work factoryapi.Work, params factoryapi.ListWorkParams) bool {
-	if params.TraceId == nil || *params.TraceId == "" {
-		return true
-	}
-	traceID := string(*params.TraceId)
-	return stringValue(work.TraceId) == traceID || stringValue(work.CurrentChainingTraceId) == traceID
-}
-
-func (s *Server) GetWork(w http.ResponseWriter, r *http.Request, id factoryapi.WorkOrTokenID) {
-	s.getWork(w, r, id, s.runtime.GetEngineStateSnapshot)
-}
-
-func (s *Server) GetWorkBySessionId(w http.ResponseWriter, r *http.Request, sessionID factoryapi.SessionID, id factoryapi.WorkOrTokenID) {
-	sessionRuntime, ok := s.requireSessionRuntime(w)
-	if !ok {
-		return
-	}
-	s.getWork(w, r, id, func(ctx context.Context) (*interfaces.EngineStateSnapshot[petri.MarkingSnapshot, *state.Net], error) {
-		return sessionRuntime.GetEngineStateSnapshotForSession(ctx, string(sessionID))
-	})
-}
-
-func (s *Server) getWork(
-	w http.ResponseWriter,
-	r *http.Request,
-	id factoryapi.WorkOrTokenID,
-	loadSnapshot func(context.Context) (*interfaces.EngineStateSnapshot[petri.MarkingSnapshot, *state.Net], error),
-) {
-	snapshot, err := loadSnapshot(r.Context())
-	if err != nil {
-		if errors.Is(err, apisurface.ErrFactorySessionNotFound) {
-			s.writeError(w, http.StatusNotFound, "factory session not found", "NOT_FOUND")
-			return
-		}
-		s.logger.Error("get engine state snapshot failed", zap.Error(err))
-		s.writeError(w, http.StatusInternalServerError, "failed to get engine state snapshot", "INTERNAL_ERROR")
-		return
-	}
-
-	token, ok := findPublicWorkToken(snapshot.Marking.Tokens, string(id))
-	if !ok {
-		s.writeError(w, http.StatusNotFound, "work not found", "NOT_FOUND")
-		return
-	}
-
-	workNamesByID := publicWorkNamesByID(snapshot.Marking.Tokens)
-	work := tokenToWork(token, snapshot.Topology)
-	work.Relations = generatedWorkRelations(token, work.Name, workNamesByID)
-	s.writeJSON(w, http.StatusOK, work)
-}
-
-func findPublicWorkToken(tokens map[string]*interfaces.Token, id string) (*interfaces.Token, bool) {
-	if token, ok := tokens[id]; ok && publicWorkToken(token) {
-		return token, true
-	}
-	for _, token := range tokens {
-		if !publicWorkToken(token) {
-			continue
-		}
-		if token.Color.WorkID == id {
-			return token, true
-		}
-	}
-	return nil, false
-}
-
 // GetStatus handles GET /status as the supported runtime status read model.
 func (s *Server) GetStatus(w http.ResponseWriter, r *http.Request) {
 	s.getStatus(w, r, s.runtime.GetEngineStateSnapshot)
@@ -1130,88 +567,6 @@ func tokenToResponse(t *interfaces.Token, includeHistory bool) factoryapi.TokenR
 	return resp
 }
 
-func tokenToWork(t *interfaces.Token, net *state.Net) factoryapi.Work {
-	name := firstNonEmptyString(t.Color.Name, t.Color.WorkID, t.ID)
-	return factoryapi.Work{
-		Name:                     name,
-		WorkId:                   stringPtrIfNotEmpty(t.Color.WorkID),
-		WorkTypeName:             stringPtrIfNotEmpty(t.Color.WorkTypeID),
-		State:                    workStateForToken(t, net),
-		ChainingTraceDepth:       intPtrIfPositive(t.Color.ChainingTraceDepth),
-		CurrentChainingTraceId:   stringPtrIfNotEmpty(firstNonEmptyString(t.Color.CurrentChainingTraceID, t.Color.TraceID)),
-		PreviousChainingTraceIds: stringSlicePtrCopy(t.Color.PreviousChainingTraceIDs),
-		TraceId:                  stringPtrIfNotEmpty(t.Color.TraceID),
-		Content:                  domainWorkContentToGeneratedPtr(t.Color.Content),
-		Tags:                     stringMapPtr(t.Color.Tags),
-	}
-}
-
-func publicWorkNamesByID(tokens map[string]*interfaces.Token) map[string]string {
-	names := make(map[string]string, len(tokens))
-	for _, token := range tokens {
-		if !publicWorkToken(token) || token.Color.WorkID == "" {
-			continue
-		}
-		names[token.Color.WorkID] = firstNonEmptyString(token.Color.Name, token.Color.WorkID, token.ID)
-	}
-	return names
-}
-
-func generatedWorkRelations(token *interfaces.Token, sourceWorkName string, workNamesByID map[string]string) *[]factoryapi.Relation {
-	if token == nil || len(token.Color.Relations) == 0 {
-		return nil
-	}
-
-	relations := make([]factoryapi.Relation, 0, len(token.Color.Relations))
-	for _, relation := range token.Color.Relations {
-		targetWorkName := firstNonEmptyString(workNamesByID[relation.TargetWorkID], relation.TargetWorkID)
-		relations = append(relations, factoryapi.Relation{
-			Type:           factoryapi.RelationType(relation.Type),
-			SourceWorkName: sourceWorkName,
-			TargetWorkName: targetWorkName,
-			TargetWorkId:   stringPtrIfNotEmpty(relation.TargetWorkID),
-			RequiredState:  stringPtrIfNotEmpty(relation.RequiredState),
-		})
-	}
-	return &relations
-}
-
-func workStateForToken(t *interfaces.Token, net *state.Net) *factoryapi.WorkState {
-	if t == nil {
-		return nil
-	}
-	workTypeID, stateName := state.SplitPlaceID(t.PlaceID)
-	if t.Color.WorkTypeID != "" {
-		workTypeID = t.Color.WorkTypeID
-	}
-	if net != nil {
-		if place, ok := net.Places[t.PlaceID]; ok {
-			workTypeID = place.TypeID
-			stateName = place.State
-		}
-	}
-	if stateName == "" {
-		return nil
-	}
-	return &factoryapi.WorkState{
-		Name: stateName,
-		Type: factoryapi.WorkStateType(state.CategoryForState(workTypesFromNet(net), workTypeID, stateName)),
-	}
-}
-
-func workTypesFromNet(net *state.Net) map[string]*state.WorkType {
-	if net == nil {
-		return nil
-	}
-	return net.WorkTypes
-}
-
-func publicWorkToken(token *interfaces.Token) bool {
-	return token != nil &&
-		token.Color.DataType != interfaces.DataTypeResource &&
-		!interfaces.IsSystemTimeToken(token)
-}
-
 func statusFromEngineStateSnapshot(snapshot interfaces.EngineStateSnapshot[petri.MarkingSnapshot, *state.Net]) factoryapi.StatusResponse {
 	categories, resources := categorizeStatusTokens(&snapshot.Marking, snapshot.Topology)
 	return factoryapi.StatusResponse{
@@ -1330,13 +685,6 @@ func stringValue(value *string) string {
 	return optional.StringValue(value)
 }
 
-func generatedWorkStateName(value *factoryapi.WorkState) string {
-	if value == nil {
-		return ""
-	}
-	return value.Name
-}
-
 func intValue(value *int) int {
 	return optional.IntValue(value)
 }
@@ -1380,149 +728,6 @@ func firstNonEmptyString(values ...string) string {
 
 func generatedStringMap(values *factoryapi.StringMap) map[string]string {
 	return optional.StringMapValue(values)
-}
-
-func generatedSubmitRelations(values *[]factoryapi.SubmitRelation) []interfaces.Relation {
-	if values == nil || len(*values) == 0 {
-		return nil
-	}
-	relations := make([]interfaces.Relation, 0, len(*values))
-	for _, relation := range *values {
-		relations = append(relations, interfaces.Relation{
-			Type:          interfaces.RelationType(relation.Type),
-			TargetWorkID:  relation.TargetWorkId,
-			RequiredState: stringValue(relation.RequiredState),
-		})
-	}
-	return relations
-}
-
-func generatedWorkRequestToDomain(req factoryapi.WorkRequest) (interfaces.WorkRequest, error) {
-	workRequest := interfaces.WorkRequest{
-		RequestID:              req.RequestId,
-		CurrentChainingTraceID: stringValue(req.CurrentChainingTraceId),
-		Type:                   interfaces.WorkRequestType(req.Type),
-	}
-	if req.Works != nil {
-		workRequest.Works = make([]interfaces.Work, 0, len(*req.Works))
-		for i, work := range *req.Works {
-			if err := validateGeneratedWorkContentAtPath(work.Content, fmt.Sprintf("works[%d].content", i)); err != nil {
-				return interfaces.WorkRequest{}, err
-			}
-			workRequest.Works = append(workRequest.Works, interfaces.Work{
-				Name:                     work.Name,
-				WorkID:                   stringValue(work.WorkId),
-				RequestID:                stringValue(work.RequestId),
-				WorkTypeID:               stringValue(work.WorkTypeName),
-				State:                    generatedWorkStateName(work.State),
-				ChainingTraceDepth:       intValue(work.ChainingTraceDepth),
-				CurrentChainingTraceID:   stringValue(work.CurrentChainingTraceId),
-				PreviousChainingTraceIDs: stringSliceValue(work.PreviousChainingTraceIds),
-				TraceID:                  stringValue(work.TraceId),
-				Content:                  workcontent.PartsFromGenerated(work.Content),
-				Payload:                  work.Payload,
-				Tags:                     generatedStringMap(work.Tags),
-			})
-		}
-	}
-	if req.Relations != nil {
-		workRequest.Relations = make([]interfaces.WorkRelation, 0, len(*req.Relations))
-		for _, relation := range *req.Relations {
-			workRequest.Relations = append(workRequest.Relations, interfaces.WorkRelation{
-				Type:           interfaces.WorkRelationType(relation.Type),
-				SourceWorkName: relation.SourceWorkName,
-				TargetWorkName: relation.TargetWorkName,
-				RequiredState:  stringValue(relation.RequiredState),
-			})
-		}
-	}
-	return workRequest, nil
-}
-
-func domainWorkContentToGeneratedPtr(parts []interfaces.WorkContentPart) *factoryapi.WorkContent {
-	return workcontent.GeneratedPtrFromParts(parts)
-}
-
-func validateGeneratedWorkContentAtPath(content *factoryapi.WorkContent, fieldPath string) error {
-	if content == nil || len(*content) == 0 {
-		return nil
-	}
-
-	for i, part := range *content {
-		pathPrefix := fmt.Sprintf("%s[%d].", fieldPath, i)
-		if _, ok := workcontent.PartFromGenerated(part); ok {
-			continue
-		}
-
-		return requestFieldValidationError{message: fmt.Sprintf("%stype must be one of text, image, TEXT, IMAGE, AUDIO, JSON, or BINARY", pathPrefix)}
-	}
-	return nil
-}
-
-func decodeSubmitWorkRequestBody(body io.Reader) (factoryapi.SubmitWorkJSONRequestBody, error) {
-	data, err := io.ReadAll(body)
-	if err != nil {
-		return factoryapi.SubmitWorkJSONRequestBody{}, err
-	}
-
-	var req factoryapi.SubmitWorkJSONRequestBody
-	if err := json.Unmarshal(data, &req); err != nil {
-		return factoryapi.SubmitWorkJSONRequestBody{}, err
-	}
-	if err := validateCanonicalWorkRequestJSONForAPI(data); err != nil {
-		return factoryapi.SubmitWorkJSONRequestBody{}, err
-	}
-
-	var fields map[string]json.RawMessage
-	if err := json.Unmarshal(data, &fields); err != nil {
-		return factoryapi.SubmitWorkJSONRequestBody{}, err
-	}
-	if err := validateSubmitWorkStructuredInputFields(fields); err != nil {
-		return factoryapi.SubmitWorkJSONRequestBody{}, err
-	}
-	if err := validateWorkContentField(fields, ""); err != nil {
-		return factoryapi.SubmitWorkJSONRequestBody{}, err
-	}
-	if strings.TrimSpace(req.Name) == "" {
-		return factoryapi.SubmitWorkJSONRequestBody{}, requestFieldValidationError{message: "name is required"}
-	}
-	return req, nil
-}
-
-func decodeWorkRequestBody(body io.Reader) (factoryapi.UpsertWorkRequestJSONRequestBody, error) {
-	data, err := io.ReadAll(body)
-	if err != nil {
-		return factoryapi.UpsertWorkRequestJSONRequestBody{}, err
-	}
-
-	var req factoryapi.UpsertWorkRequestJSONRequestBody
-	if err := json.Unmarshal(data, &req); err != nil {
-		return factoryapi.UpsertWorkRequestJSONRequestBody{}, err
-	}
-	if err := validateCanonicalWorkRequestJSONForAPI(data); err != nil {
-		return factoryapi.UpsertWorkRequestJSONRequestBody{}, err
-	}
-
-	if req.Works == nil || len(*req.Works) == 0 {
-		return req, nil
-	}
-
-	var rawRequest struct {
-		Works []map[string]json.RawMessage `json:"works"`
-	}
-	if err := json.Unmarshal(data, &rawRequest); err != nil {
-		return factoryapi.UpsertWorkRequestJSONRequestBody{}, err
-	}
-
-	for i := range *req.Works {
-		if i >= len(rawRequest.Works) {
-			return req, nil
-		}
-		if err := validateWorkContentField(rawRequest.Works[i], fmt.Sprintf("works[%d].", i)); err != nil {
-			return factoryapi.UpsertWorkRequestJSONRequestBody{}, err
-		}
-	}
-	return req, nil
 }
 
 func decodeNamedFactoryBody(body io.Reader) (factoryapi.Factory, error) {
@@ -1710,74 +915,4 @@ func derefGeneratedWorkContent(content *factoryapi.WorkContent) factoryapi.WorkC
 		return nil
 	}
 	return *content
-}
-
-func applyStableTraceToWorkRequest(req *interfaces.WorkRequest) {
-	if req == nil || len(req.Works) == 0 {
-		return
-	}
-	traceID := ""
-	if req.CurrentChainingTraceID != "" {
-		traceID = req.CurrentChainingTraceID
-	}
-	if traceID == "" {
-		for _, work := range req.Works {
-			if work.CurrentChainingTraceID != "" {
-				traceID = work.CurrentChainingTraceID
-				break
-			}
-			if work.TraceID != "" {
-				traceID = work.TraceID
-				break
-			}
-		}
-	}
-	if traceID == "" {
-		traceID = "trace-" + req.RequestID
-	}
-	if req.CurrentChainingTraceID == "" {
-		req.CurrentChainingTraceID = traceID
-	}
-	for i := range req.Works {
-		if req.Works[i].CurrentChainingTraceID == "" {
-			if req.Works[i].TraceID != "" {
-				req.Works[i].CurrentChainingTraceID = req.Works[i].TraceID
-			} else {
-				req.Works[i].CurrentChainingTraceID = traceID
-			}
-		}
-		if req.Works[i].TraceID == "" {
-			req.Works[i].TraceID = req.Works[i].CurrentChainingTraceID
-		}
-	}
-}
-
-func generatedPayloadToRawMessage(payload any) (json.RawMessage, error) {
-	if payload == nil {
-		return nil, nil
-	}
-	return json.Marshal(payload)
-}
-
-func submitWorkBadRequestMessage(err error) (string, bool) {
-	if err == nil {
-		return "", false
-	}
-	message := err.Error()
-	if strings.HasPrefix(message, "work_request:") {
-		return submitWorkTypeNameMessage(message), true
-	}
-	if strings.Contains(message, "unknown work type") || strings.Contains(message, "work type") && strings.Contains(message, "not found") {
-		return submitWorkTypeNameMessage(message), true
-	}
-	return "", false
-}
-
-func submitWorkTypeNameMessage(message string) string {
-	message = strings.ReplaceAll(message, "work_type_name", "workTypeName")
-	message = strings.ReplaceAll(message, "work_type_id", "workTypeName")
-	if strings.Contains(message, "work type name") {
-		return message
-	}
-	return strings.ReplaceAll(message, "work type", "work type name")
 }
